refactor(decouple): keep MCO-only agent files in mongodb-community-operator/pkg/agent
Move replica_set_port_manager, agenthealth back into the MCO boundary; delete
agentflags (unused dead code). pkg/agent now contains only agent_readiness, which
is genuinely shared between the MCO and Enterprise AppDB controllers.

replica_set_port_manager imports pkg/agent as plain "agent" for the PodState type;
replica_set_controller uses mcoagent alias for the MCO package alongside the shared
agent import for readiness calls.
