+++ mongodb-community-operator/cmd/versionhook/main.go
@@ -16,7 +16,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/mongodb/mongodb-kubernetes/pkg/agent"
+	"github.com/mongodb/mongodb-kubernetes/mongodb-community-operator/pkg/agent"
 )
 
 const (
